weather/pyproject.toml
[project]
name = "weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.1",
    "fastmcp",
    "python-dotenv",
    "aiohttp",
    "uvicorn",
]
